LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/select-add-optgroup-expected.txt



PASS select.add() with an index should work when the target is inside an optgroup.
